requirements.txt
avro==1.11.3
certifi==2025.4.26
charset-normalizer==3.4.2
confluent-kafka==2.3.0
fastavro==1.11.1
idna==3.10
python-dotenv==1.0.0
requests==2.32.3
urllib3==2.4.0
uuid==1.30
